+++ src/prepare/prepare-entries.ts
@@ -1,0 +1,63 @@
+import { existsSync } from 'fs'
+import { glob } from 'glob'
+import { BINARY_TAG, availableExtensions } from '../constants'
+
+import { collectSourceEntriesByExportPath } from '../entries'
+import { sourceFilenameToExportFullPath } from '../utils'
+
+// For `prepare` command
+export async function collectSourceEntries(sourceFolderPath: string) {
+  const bins = new Map<string, string>()
+  const exportsEntries = new Map<string, Record<string, string>>()
+  if (!existsSync(sourceFolderPath)) {
+    return {
+      bins,
+      exportsEntries,
+    }
+  }
+
+  // Match with global patterns
+  // bin/**/*.<ext>, bin/**/index.<ext>
+  const binPattern = `bin/**/*.{${[...availableExtensions].join(',')}}`
+  const srcPattern = `**/*.{${[...availableExtensions].join(',')}}`
+
+  const binMatches = await glob(binPattern, {
+    cwd: sourceFolderPath,
+    nodir: true,
+  })
+
+  const srcMatches = await glob(srcPattern, {
+    cwd: sourceFolderPath,
+    nodir: true,
+  })
+
+  for (const file of binMatches) {
+    // convert relative path to export path
+    const exportPath = sourceFilenameToExportFullPath(file)
+    await collectSourceEntriesByExportPath(
+      sourceFolderPath,
+      exportPath,
+      bins,
+      exportsEntries,
+    )
+  }
+
+  for (const file of srcMatches) {
+    const binExportPath = file
+      .replace(/^bin/, BINARY_TAG)
+      // Remove index.<ext> to [^index].<ext> to build the export path
+      .replace(/(\/index)?\.[^/]+$/, '')
+
+    await collectSourceEntriesByExportPath(
+      sourceFolderPath,
+      binExportPath,
+      bins,
+      exportsEntries,
+    )
+  }
+
+  return {
+    bins,
+    exportsEntries,
+  }
+}
